Extract CacheStatsAssertions for reuse
Extract `CacheStatsAssertions` from `TestCachingJdbcClient`.
